requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cssselect2==0.7.0
Django==5.1.1
django-cors-headers==4.6.0
django-extensions==4.1
django-filter==25.1
django-formtools==2.5.1
django-otp==1.6.0
django-phonenumber-field==8.1.0
django-two-factor-auth==1.17.0
django_paystack==1.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
fonttools==4.54.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
phonenumbers==9.0.4
pillow==10.4.0
psycopg==3.2.6
psycopg-binary==3.2.6
pycparser==2.22
pydyf==0.11.0
pygame==2.6.1
PyJWT==2.9.0
pyotp==2.9.0
pypaystack2==2.0.3
pyphen==0.15.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-decouple==3.8
qrcode==7.4.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.2
urllib3==2.2.3
weasyprint==63.0
webencodings==0.5.1
zopfli==0.2.3.post1
